CORE-4355: Move packaging format version 1 classes into new package
- Move packaging format version 1 classes into their own package
- Add V1 to format version 1 class names
