+++ libs/packaging/packaging/src/main/kotlin/net/corda/libs/packaging/internal/v1/CpiLoaderV1.kt
@@ -1,4 +1,4 @@
-package net.corda.libs.packaging.internal
+package net.corda.libs.packaging.internal.v1
 
 import net.corda.libs.packaging.Cpi
 import net.corda.libs.packaging.Cpk
@@ -11,6 +11,7 @@ import net.corda.libs.packaging.core.CpkMetadata
 import net.corda.libs.packaging.core.exception.PackagingException
 import net.corda.libs.packaging.PackagingConstants.CPI_GROUP_POLICY_ENTRY
 import net.corda.libs.packaging.certSummaryHash
+import net.corda.libs.packaging.internal.CpiImpl
 import net.corda.v5.crypto.DigestAlgorithmName
 import net.corda.v5.crypto.SecureHash
 import java.io.InputStream
@@ -23,7 +24,7 @@ import java.util.jar.JarInputStream
 import java.util.jar.Manifest
 import java.util.zip.ZipEntry
 
-internal object CpiLoader {
+internal object CpiLoaderV1 {
     private const val cpkFileExtension = ".cpk"
     private fun isCpk(entry : ZipEntry) = !entry.isDirectory && entry.name.endsWith(cpkFileExtension)
     private fun isGroupPolicy(entry : ZipEntry) = !entry.isDirectory && entry.name.endsWith(CPI_GROUP_POLICY_ENTRY)
@@ -68,7 +69,7 @@ internal object CpiLoader {
                             cpks += cpk
                             cpkMetadata += cpk.metadata
                         } else {
-                            cpkMetadata += CpkLoader.loadMetadata(uncloseableInputStream,
+                            cpkMetadata += CpkLoaderV1.loadMetadata(uncloseableInputStream,
                                 cpkLocation = cpiLocation?.plus("/${entry.name}"),
                                 verifySignature = verifySignature)
                         }
