+++ libs/packaging/packaging/src/main/kotlin/net/corda/libs/packaging/internal/v1/CpkLoaderV1.kt
@@ -1,4 +1,4 @@
-package net.corda.libs.packaging.internal
+package net.corda.libs.packaging.internal.v1
 
 import net.corda.libs.packaging.Cpk
 import net.corda.libs.packaging.CpkDocumentReader
@@ -17,6 +17,8 @@ import net.corda.libs.packaging.core.exception.CordappManifestException
 import net.corda.libs.packaging.core.exception.InvalidSignatureException
 import net.corda.libs.packaging.core.exception.LibraryIntegrityException
 import net.corda.libs.packaging.core.exception.PackagingException
+import net.corda.libs.packaging.internal.CpkImpl
+import net.corda.libs.packaging.internal.FormatVersionReader
 import net.corda.v5.base.util.loggerFor
 import net.corda.v5.crypto.DigestAlgorithmName
 import net.corda.v5.crypto.SecureHash
@@ -39,8 +41,8 @@ import java.util.jar.JarInputStream
 import java.util.jar.Manifest
 import java.util.zip.ZipEntry
 
-internal object CpkLoader {
-    private val logger = loggerFor<CpkLoader>()
+internal object CpkLoaderV1 {
+    private val logger = loggerFor<CpkLoaderV1>()
 
     private const val CPK_TYPE = "Corda-CPK-Type"
