+++ libs/packaging/packaging/src/main/kotlin/net/corda/libs/packaging/internal/v1/UncloseableInputStream.kt
@@ -1,4 +1,4 @@
-package net.corda.libs.packaging.internal
+package net.corda.libs.packaging.internal.v1
 
 import java.io.FilterInputStream
 import java.io.InputStream
